SmartTrainingScheme/backend/requirements.txt
Django>=4.2
djangorestframework
networkx
pandas
numpy
openpyxl
tqdm
